apply review: WeatherImageを別ファイルに切り出した
